Name udev rule according to systemd expectations
After installing `libfobos` on Fedora 40, I had troubles getting
correct permissions to access FobosSDR device.

According to https://wiki.archlinux.org/title/udev and
https://github.com/systemd/systemd/issues/4288#issuecomment-348166161
"For any rule adding the uaccess tag to be effective, the name of the
file it is defined in has to lexically precede.". So lets rename the
udev rule accordingly.